refactor: move stage type to enums
