+++ backend/app/models/enums.py
@@ -7,3 +7,14 @@ class StatusType(str, enum.Enum):
     FINISHED = "finished"
     POSTPONED = "postponed"
     CANCELLED = "cancelled"
+
+
+class StageType(str, enum.Enum):
+    GROUP = "group"
+    ROUND_OF_32 = "round_of_32"
+    ROUND_OF_16 = "round_of_16"
+    QUARTER_FINAL = "quarter_final"
+    SEMI_FINAL = "semi_final"
+    THIRD_PLACE = "third_place"
+    FINAL = "final"
+    OTHER = "other"
