+++ backend/tests/integration/test_get_ranked_tournament_teams.py
@@ -3,7 +3,7 @@ from datetime import date
 import pytest
 
 from app.api.v1.services import tournament_teams as tournament_teams_service
-from app.models.match import StageType
+from app.models.enums import StageType
 from app.models.team import Team
 from app.models.tournament import Tournament
 from app.models.tournament_team import TournamentTeam
