+++ backend/tests/integration/test_refresh_team_rankings.py
@@ -2,8 +2,8 @@ from datetime import date
 
 import pytest
 
-from app.models.enums import StatusType
-from app.models.match import Match, StageType
+from app.models.enums import StageType, StatusType
+from app.models.match import Match
 from app.models.standing import Standing
 from app.models.team import Team
 from app.models.tournament import Tournament
