+++ backend/tests/unit/utils/test_match_events_cache_helper.py
@@ -10,8 +10,7 @@ from app.constants.cache_ttl import (
     MATCH_EVENTS_LIVE_KNOCKOUT_TTL,
     MATCH_EVENTS_SOON_SCHEDULED_TTL,
 )
-from app.models.enums import StatusType
-from app.models.match import StageType
+from app.models.enums import StageType, StatusType
 
 
 def test_get_match_events_ttl_returns_group_live_ttl():
